refactor(rel-import): progressindicator, sidenav updated
