+++ src/components/SideNav/index.js
@@ -1,1 +1,14 @@
 export default from './SideNav';
+export {
+  SideNavDetails,
+  SideNavFooter,
+  SideNavHeader,
+  SideNavIcon,
+  SideNavItem,
+  SideNavItems,
+  SideNavLink,
+  SideNavLinkText,
+  SideNavMenu,
+  SideNavMenuItem,
+  SideNavSwitcher,
+} from 'carbon-components-react';
